Update ArduinoFlatBox.ino
This is actually changes on top of other changes by other people.  This version supports the ESP32 board and used Bluetooth to communicate with the box, so I have no USB cable leaving the pier, as my panel is mounted to the observatory wall for use with multiple telescopes.  As such, it does not support any shutter commands.